+++ packages/zosfiles/__tests__/__system__/methods/copy/Copy.system.test.ts
@@ -19,11 +19,11 @@ import { ITestPropertiesSchema } from "../../../../../../__tests__/__src__/prope
 import { join } from "path";
 import { readFileSync } from "fs";
 import { ITestEnvironment } from "../../../../../../__tests__/__src__/environment/ITestEnvironment";
-import { tmpdir } from "os";
 import path = require("path");
 import * as fs from "fs";
+import * as util from "util";
+import { tmpdir } from "os";
 import { ZosmfRestClient } from "@zowe/core-for-zowe-sdk";
-
 
 let REAL_SESSION: Session;
 let REAL_TARGET_SESSION: Session;
@@ -65,6 +65,14 @@ describe("Copy", () => {
             }
         });
         describe("Success Scenarios", () => {
+            afterEach(async () => {
+                try {
+                    await Delete.dataSet(REAL_SESSION, fromDataSetName);
+                    await Delete.dataSet(REAL_SESSION, toDataSetName);
+                } catch (err) {
+                    Imperative.console.info(`Error: ${inspect(err)}`);
+                }
+            });
             describe("Sequential > Sequential", () => {
                 beforeEach(async () => {
                     try {
@@ -106,7 +114,7 @@ describe("Copy", () => {
                     expect(contents1.toString()).toEqual(contents2.toString());
                 });
             });
-            describe("Partioned > Partioned", () => {
+            describe("Partitioned > Partitioned", () => {
                 let downloadDir: string;
                 beforeEach(async () => {
                     try {
@@ -127,8 +135,6 @@ describe("Copy", () => {
                 it("Should copy a partitioned data set", async () => {
                     let error;
                     let response;
-                    let contents1;
-                    let contents2;
 
                     try {
                         response = await Copy.dataSet(
@@ -138,8 +144,6 @@ describe("Copy", () => {
                                 dsn:fromDataSetName
                             }}
                         );
-                        contents1 = await Get.dataSet(REAL_SESSION, fromDataSetName);
-                        contents2 = await Get.dataSet(REAL_SESSION, toDataSetName);
                         Imperative.console.info(`Response: ${inspect(response)}`);
                     } catch (err) {
                         error = err;
@@ -151,13 +155,42 @@ describe("Copy", () => {
                     expect(response).toBeTruthy();
                     expect(response.success).toBe(true);
                     expect(response.commandResponse).toContain(ZosFilesMessages.datasetCopiedSuccessfully.message);
-
-                    expect(contents1).toBeTruthy();
-                    expect(contents2).toBeTruthy();
-                    expect(contents1.toString()).toEqual(contents2.toString());
                 });
-                afterEach(() => {
-                    fs.rmSync(downloadDir, { recursive: true, force: true });
+                it("Should handle truncation errors and log them to a file", async () => {
+                    let error;
+                    let response;
+
+                    const uploadFileToDatasetSpy = jest.spyOn(Upload, 'fileToDataset').mockImplementation(async (session, filePath, dsn) => {
+                        if (filePath === fileLocation) {
+                            throw new Error("Truncation of a record occurred during an I/O operation");
+                        }
+                        return Promise.resolve() as any;
+                    });
+                    const copyDataSetSpy = jest.spyOn(Copy, 'dataSet').mockImplementation(async () => {
+                        return {
+                            success: true,
+                            commandResponse: ZosFilesMessages.datasetCopiedSuccessfully.message + " " +
+                            util.format(ZosFilesMessages.membersContentTruncated.message)
+                        };
+                    });
+                    try {
+                        response = await Copy.dataSet(
+                            REAL_SESSION,
+                            {dsn: toDataSetName},
+                            {"from-dataset": {
+                                dsn:fromDataSetName
+                            }}
+                        );
+                    } catch (err) {
+                        error = err;
+                        Imperative.console.info(`Error: ${inspect(err)}`);
+                    }
+                    expect(response).toBeTruthy();
+                    expect(response.success).toBe(true);
+                    expect(response.commandResponse).toContain(ZosFilesMessages.datasetCopiedSuccessfully.message + " " +
+                        util.format(ZosFilesMessages.membersContentTruncated.message));
+                    uploadFileToDatasetSpy.mockRestore();
+                    copyDataSetSpy.mockRestore();
                 });
             });
             describe("Member > Member", () => {
@@ -594,6 +627,45 @@ describe("Copy", () => {
                 expect(contents2).toBeTruthy();
                 expect(contents1.toString()).toEqual(contents2.toString());
             });
+        });
+    });
+
+    describe("hasIdenticalMemberNames", () => {
+        beforeEach(async () => {
+            try {
+                await Create.dataSet(REAL_SESSION, CreateDataSetTypeEnum.DATA_SET_PARTITIONED, fromDataSetName);
+                await Create.dataSet(REAL_SESSION, CreateDataSetTypeEnum.DATA_SET_PARTITIONED, toDataSetName);
+                await Upload.fileToDataset(REAL_SESSION, fileLocation, fromDataSetName);
+                await Upload.fileToDataset(REAL_SESSION, fileLocation, toDataSetName);
+            }
+            catch (err) {
+                Imperative.console.info(`Error: ${inspect(err)}`);
+            }
+        });
+        afterEach(async () => {
+            try {
+                await Delete.dataSet(REAL_SESSION, fromDataSetName);
+                await Delete.dataSet(REAL_SESSION, toDataSetName);
+            } catch (err) {
+                Imperative.console.info(`Error: ${inspect(err)}`);
+            }
+        });
+        it("should return true if the source and target data sets have identical member names", async () => {
+            const sourceResponse = await List.allMembers(REAL_SESSION, fromDataSetName);
+            const sourceMemberList = sourceResponse.apiResponse.items.map((item: { member: any; }) => item.member);
+            const response = await Copy["hasIdenticalMemberNames"](REAL_SESSION, sourceMemberList, toDataSetName);
+            expect(response).toBe(true);
+        });
+
+        it("should return false if the source and target data sets do not have identical member names", async () => {
+            await Delete.dataSet(REAL_SESSION, toDataSetName);
+            await Create.dataSet(REAL_SESSION, CreateDataSetTypeEnum.DATA_SET_PARTITIONED, toDataSetName);
+
+            const sourceResponse = await List.allMembers(REAL_SESSION, fromDataSetName);
+            const sourceMemberList = sourceResponse.apiResponse.items.map((item: { member: any; }) => item.member);
+
+            const response = await Copy["hasIdenticalMemberNames"](REAL_SESSION, sourceMemberList, toDataSetName);
+            expect(response).toBe(false);
         });
     });